Moved unit tests into unit folder
