+++ api/tests/unit/middlewares/logger.test.js
@@ -1,4 +1,4 @@
-const { logger } = require('../../src/middlewares');
+const { logger } = require('../../../src/middlewares');
 
 describe('Test logger middleware', () => {
   test('Should be a function', () => {
