+++ api/tests/unit/utils/debug.test.js
@@ -1,4 +1,4 @@
-const debug = require('../../src/utils/debug')('env-test');
+const debug = require('../../../src/utils/debug')('env-test');
 
 describe('Test debugging utility', () => {
   test('Should be disabled in test env', () => {
